requirements.txt
asgiref==3.2.10
autopep8==1.5.3
boto3==1.14.21
botocore==1.17.21
Django==3.0.8
django-bootstrap-modal-forms==2.0.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.10.0
Pillow==7.2.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
